UX & UI Updates
Separated code into panels, playblast, & properties files
Added Shading controls
Added icons to Panels
Added Icon button to View 3D
